+++ common/model/src/main/java/is/codion/common/model/table/TableConditionModel.java
@@ -34,7 +34,7 @@ public interface TableConditionModel<C> {
 	/**
 	 * @return an unmodifiable map containing the condition models available in this table condition model, mapped to their respective column identifiers
 	 */
-	Map<C, ColumnConditionModel<? extends C, ?>> conditionModels();
+	Map<C, ColumnConditionModel<C, ?>> conditionModels();
 
 	/**
 	 * The condition model associated with {@code columnIdentifier}
@@ -43,7 +43,7 @@ public interface TableConditionModel<C> {
 	 * @return the {@link ColumnConditionModel} for the {@code columnIdentifier}
 	 * @throws IllegalArgumentException in case no condition model exists for the given columnIdentifier
 	 */
-	<T> ColumnConditionModel<? extends C, T> conditionModel(C columnIdentifier);
+	<T> ColumnConditionModel<C, T> conditionModel(C columnIdentifier);
 
 	/**
 	 * Clears the search state of all the condition models, disables them and
@@ -73,7 +73,7 @@ public interface TableConditionModel<C> {
 	 * @param <C> the column identifier type
 	 * @return a new {@link TableConditionModel}
 	 */
-	static <C> TableConditionModel<C> tableConditionModel(Collection<ColumnConditionModel<? extends C, ?>> conditionModels) {
+	static <C> TableConditionModel<C> tableConditionModel(Collection<ColumnConditionModel<C, ?>> conditionModels) {
 		return new DefaultTableConditionModel<>(requireNonNull(conditionModels));
 	}
 }