+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityConditionModelFactory.java
@@ -44,7 +44,7 @@ public class SwingEntityConditionModelFactory extends EntityConditionModelFactor
 	}
 
 	@Override
-	public Optional<ColumnConditionModel<? extends Attribute<?>, ?>> createConditionModel(Attribute<?> attribute) {
+	public Optional<ColumnConditionModel<Attribute<?>, ?>> createConditionModel(Attribute<?> attribute) {
 		if (attribute instanceof ForeignKey) {
 			ForeignKey foreignKey = (ForeignKey) attribute;
 			if (definition(foreignKey.referencedType()).smallDataset()) {